Beginning of code to accommodate two new file types (raw SIF data and surface area data)
Deprecated as data was proved to be faulty.

This also contains code to calculate radius of circle based on
circumference points, in an effort to use Nejati’s thesis formula to
impose maximum threshold on K+G values, in an effort to eliminate
outliers. However, a lower threshold of 10 * Young’s Modulus (eg. 100
GPa or 1e+11) was more effective, so this code was deprecated.